Intro to codechef
I have given a bried intro to codechef and types of competition and ranking system here. I would like to arrange the folder of codechef files to arrange them in such  a manner that div 1 div 2 and div 3 questions should be separated as level of questions is really different for them.